# openapi-test/src/openapi.yaml
openapi: 3.0.2
info:
  title: FastAPI
  version: 0.1.0
  description: "MyGPT backend API test"
  contact:
    name: HongYu Xiao
    email: hongyuxiao05@gmail.com
servers:
- url: "http://localhost:5051"
  description: "Development environment"
- url: "https://cloneapi.mygpt.fi"
  description: "Production environment"
paths:
...
  /brains/:
    get:
      tags:
        - Brain
      summary: Brain Endpoint
      description: >-
        Retrieve all brains for the current user.


        - `current_user`: The current authenticated user.

        - Returns a list of all brains registered for the user.


        This endpoint retrieves all the brains associated with the current
        authenticated user. It returns a list of brains objects

        containing the brain ID and brain name for each brain.
      operationId: brain_endpoint_brains__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
        - AuthBearer: []
        - AuthBearer: []
    post:
      tags:
        - Brain
      summary: Create Brain Endpoint
      description: >-
        Create a new brain with given
            name
            status
            model
            max_tokens
            temperature
        In the brains table & in the brains_users table and put the creator user
        as 'Owner'
      operationId: create_brain_endpoint_brains__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateBrainProperties'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - AuthBearer: []
        - AuthBearer: []
  /brains/default/:
    get:
      tags:
        - Brain
      summary: Get Default Brain Endpoint
      description: >-
        Retrieve the default brain for the current user. If the user doesnt have
        one, it creates one.


        - `current_user`: The current authenticated user.

        - Returns the default brain for the user.


        This endpoint retrieves the default brain associated with the current
        authenticated user.

        The default brain is defined as the brain marked as default in the
        brains_users table.
      operationId: get_default_brain_endpoint_brains_default__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
      security:
        - AuthBearer: []
        - AuthBearer: []
  /brains/{brain_id}/:
    get:
      tags:
        - Brain
      summary: Get Brain Endpoint
      description: >-
        Retrieve details of a specific brain by brain ID.


        - `brain_id`: The ID of the brain to retrieve details for.

        - Returns the brain ID and its history.


        This endpoint retrieves the details of a specific brain identified by
        the provided brain ID. It returns the brain ID and its

        history, which includes the brain messages exchanged in the brain.
      operationId: get_brain_endpoint_brains__brain_id___get
      parameters:
        - required: true
          schema:
            title: Brain Id
            type: string
            format: uuid
          name: brain_id
          in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - AuthBearer: []
        - AuthBearer: []
    put:
      tags:
        - Brain
      summary: Update Brain Endpoint
      description: Update an existing brain with new brain configuration
      operationId: update_brain_endpoint_brains__brain_id___put
      parameters:
        - required: true
          schema:
            title: Brain Id
            type: string
            format: uuid
          name: brain_id
          in: path
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BrainUpdatableProperties'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - AuthBearer: []
        - AuthBearer: []
  /brains/{brain_id}/default:
    post:
      tags:
        - Brain
      summary: Set As Default Brain Endpoint
      description: Set a brain as default for the current user.
      operationId: set_as_default_brain_endpoint_brains__brain_id__default_post
      parameters:
        - required: true
          schema:
            title: Brain Id
            type: string
            format: uuid
          name: brain_id
          in: path
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - AuthBearer: []
        - AuthBearer: []
...
  /crawl:
    post:
      tags:
        - Crawl
      summary: Crawl Endpoint
      description: Crawl a website and process the crawled data.
      operationId: crawl_endpoint_crawl_post
      parameters:
        - description: The ID of the brain
          required: true
          schema:
            title: Brain Id
            type: string
            description: The ID of the brain
            format: uuid
          name: brain_id
          in: query
        - required: false
          schema:
            title: Enable Summarization
            type: boolean
            default: false
          name: enable_summarization
          in: query
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CrawlWebsite'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - AuthBearer: []
        - AuthBearer: []
  /crawl/linkedin:
    post:
      tags:
        - Crawl
      summary: Crawl LinkdIn Endpoint
      description: Crawl Linkdin site and process the crawled data.
      operationId: crawl_linkedin_endpoint_crawl_post
      parameters:
        - description: The ID of the brain
          required: true
          schema:
            title: Brain Id
            type: string
            description: The ID of the brain
            format: uuid
          name: brain_id
          in: query
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
        - AuthBearer: []
        - AuthBearer: []
...
components:
  schemas:
...
    BrainUpdatableProperties:
      title: BrainUpdatableProperties
      type: object
      properties:
        name:
          title: Name
          type: string
        description:
          title: Description
          type: string
        temperature:
          title: Temperature
          type: number
        model:
          title: Model
          type: string
        max_tokens:
          title: Max Tokens
          type: integer
        openai_api_key:
          title: Openai Api Key
          type: string
        status:
          title: Status
          type: string
        prompt_id:
          title: Prompt Id
          type: string
          format: uuid
...
    CrawlWebsite:
      title: CrawlWebsite
      required:
        - url
      type: object
      properties:
        url:
          title: Url
          type: string
        js:
          title: Js
          type: boolean
          default: false
        depth:
          title: Depth
          type: integer
          default: 1
        max_pages:
          title: Max Pages
          type: integer
          default: 100
        max_time:
          title: Max Time
          type: integer
          default: 60
    CreateBrainProperties:
      title: CreateBrainProperties
      type: object
      properties:
        name:
          title: Name
          type: string
          default: Default brain
        description:
          title: Description
          type: string
          default: This is a description
        status:
          title: Status
          type: string
          default: private
        model:
          title: Model
          type: string
          default: gpt-3.5-turbo
        temperature:
          title: Temperature
          type: number
          default: 0
        max_tokens:
          title: Max Tokens
          type: integer
          default: 256
        openai_api_key:
          title: Openai Api Key
          type: string
        prompt_id:
          title: Prompt Id
          type: string
          format: uuid
        linkedin:
          title: Linkedin Path
          type: string
        extraversion:
          title: Extraversion level
          type: integer
          default: 0
        neuroticism:
          title: Neuroticism level
          type: integer
          default: 0
        conscientiousness:
          title: Conscientiousness level
          type: integer
          default: 0
...
    HTTPValidationError:
      title: HTTPValidationError
      type: object
      properties:
        detail:
          title: Detail
          type: array
          items:
            $ref: '#/components/schemas/ValidationError'
...
    ValidationError:
      title: ValidationError
      required:
        - loc
        - msg
        - type
      type: object
      properties:
        loc:
          title: Location
          type: array
          items:
            anyOf:
              - type: string
              - type: integer
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
  securitySchemes:
    AuthBearer:
      type: http
      scheme: bearer